Created a material tied to a camera that updates in real time
The camera will capture whatever is in front of it, including the player, and cast it to a material. By putting the camera behind a plane and mirroring the plane it can create the effect of a mirror. This effect will be used in the tutorial area (but can be used elsewhere) as a wall blocking the player from the training range so they are forced to go to the terminal and choose their loadout. After they choose the wall will slide up. or down based on what we perfer.